backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
argon2-cffi==25.1.0
google-cloud-storage==2.19.0
